study/comment/models.py
from django.db import models
from datetime import datetime


class Comment(models.Model):
    comment = models.CharField(max_length=200)
    pub_date = models.DateTimeField('date published')

    def __unicode__(self):
        return self.question

    def was_published_today(self):
        return self.pub_date.date() == datetime.date.today()
