requirements/dev.txt
coverage>=3.7.1
django-nose>=1.4.3
flake8>=2.1.0
